Add files via upload
The code presented in this project is based on the activities and guidelines provided by FreeCodeCamp for the purpose of creating a project. While we have made efforts to implement these instructions accurately, please note that this project is not an official FreeCodeCamp-endorsed product, and it is not affiliated with FreeCodeCamp or its team.

The use of FreeCodeCamp's activities as a foundation for this project is solely for educational and skill-building purposes. Any modifications, additions, or customizations made to the original code are the result of independent work and creativity.

Although we strive to maintain code quality and adhere to best practices, there might be inadvertent errors, bugs, or vulnerabilities present in the project. Therefore, users are advised to exercise caution and review the code thoroughly before deploying it in any production or critical environment.

Furthermore, any support, maintenance, or updates provided for this project are not the responsibility of FreeCodeCamp or its team. Users should direct any inquiries or issues related to this project to the respective project maintainers and contributors.

By using this project or its code, you acknowledge that you understand and accept this disclaimer. If you disagree with the terms mentioned herein, we kindly ask you to refrain from using or accessing this project.